Add posts VuePress plugin
