+++ package.json
@@ -19,15 +19,11 @@
     "@vuepress/plugin-theme-data": "2.0.0-beta.51",
     "@vuepress/utils": "2.0.0-beta.51",
     "autoprefixer": "10.4.10",
-    "chokidar": "3.5.3",
-    "iso-639-1": "2.1.15",
     "jsonc-parser": "3.2.0",
     "markdown-it-image-lazy-loading": "1.2.0",
     "markdown-it-imsize": "2.0.1",
-    "mermaid": "9.1.6",
     "postcss": "8.4.16",
     "postcss-100vh-fix": "1.0.2",
-    "reading-time": "1.5.0",
     "shiki": "0.10.1",
     "sitemap": "7.1.1",
     "tsm": "2.2.2",
@@ -36,17 +32,17 @@
     "vue": "3.2.39",
     "vue-router": "4.1.5",
     "vue3-click-away": "1.2.4",
-    "vuepress-plugin-mermaid-js": "workspace:1.0.0",
-    "vuepress-plugin-open-graph": "workspace:1.0.0",
-    "vuepress-plugin-remove-html-extension": "workspace:1.0.0",
-    "vuepress-plugin-umami-analytics": "workspace:1.0.0",
+    "vuepress-plugin-mermaid-js": "workspace:*",
+    "vuepress-plugin-open-graph": "workspace:*",
+    "vuepress-plugin-posts": "workspace:*",
+    "vuepress-plugin-remove-html-extension": "workspace:*",
+    "vuepress-plugin-umami-analytics": "workspace:*",
     "vuepress-vite": "2.0.0-beta.51"
   },
   "devDependencies": {
     "@logux/eslint-config": "47.2.0",
     "@logux/stylelint-config": "0.12.0",
     "@size-limit/file": "8.1.0",
-    "@types/mermaid": "8.2.9",
     "@types/rss": "0.0.29",
     "@typescript-eslint/eslint-plugin": "5.37.0",
     "@typescript-eslint/parser": "5.37.0",
