+++ plugins/mermaid-js/tsconfig.json
@@ -7,5 +7,6 @@
     "rootDir": "./",
     "noEmit": false
   },
-  "include": ["./"]
+  "include": ["./**/*"],
+  "exclude": ["./dist"]
 }
