+++ plugins/posts/index.ts
@@ -1,1 +1,2 @@
 export * from './server/index.js'
+export * from './index.js'
