+++ plugins/posts/posts-shim.d.ts
@@ -1,0 +1,5 @@
+declare module '@temp/posts/data' {
+  import type { Post } from './client/index.js'
+
+  export let posts: Post[]
+}
