+++ plugins/posts/readme.md
@@ -1,0 +1,34 @@
+# vuepress-plugin-posts
+
+<img src="https://avatars.githubusercontent.com/u/48539483?s=100" align="right" alt="VuePress" />
+
+![Version](https://img.shields.io/npm/v/vuepress-plugin-posts.svg)
+![Downloads](https://img.shields.io/npm/dm/vuepress-plugin-posts.svg)
+
+VuePress plugin for blogging.
+
+## Installation
+
+```sh
+npm install --save vuepress-plugin-posts
+```
+
+## Usage
+
+Add plugin to your VuePress config:
+
+```js
+import { postsPlugin } from 'vuepress-plugin-posts'
+
+export default {
+  plugins: [postsPlugin()],
+}
+```
+
+## Contribution
+
+See [Contributing Guide](https://github.com/azat-io/azat-io/blob/main/contributing.md).
+
+## License
+
+MIT
