--- plugins/posts/typings/index.d.ts
@@ -1,12 +1,0 @@
-export interface Post {
-  availableLanguages: {
-    language: string
-    path: string
-  }[]
-  formattedDate: string
-  date: string
-  language: string
-  path: string
-  readingTime: number
-  title: string
-}
