+++ plugins/remove-html-extension/tsconfig.json
@@ -6,5 +6,6 @@
     "outDir": "./dist",
     "noEmit": false
   },
-  "include": ["./"]
+  "include": ["./**/*"],
+  "exclude": ["./dist"]
 }
