+++ plugins/umami-analytics/changelog.md
@@ -1,0 +1,7 @@
+# Change log
+
+This project adheres to [Semantic Versioning](http://semver.org/).
+
+## 0.1.0
+
+- Initial release.
