+++ vuepress.config.ts
@@ -7,6 +7,7 @@ import { openGraphPlugin } from 'vuepress-plugin-open-graph'
 import { mermaidJsPlugin } from 'vuepress-plugin-mermaid-js'
 import { viteBundler } from '@vuepress/bundler-vite'
 import { shikiPlugin } from '@vuepress/plugin-shiki'
+import { postsPlugin } from 'vuepress-plugin-posts'
 import { getDirname, path } from '@vuepress/utils'
 import { defineUserConfig } from '@vuepress/cli'
 import postcss100vhFix from 'postcss-100vh-fix'
@@ -17,7 +18,6 @@ import { loadTheme } from 'shiki'
 
 import { copyCodePlugin } from './plugins/copy-code/index.js'
 import { sitemapPlugin } from './plugins/sitemap/index.js'
-import { postsPlugin } from './plugins/posts/index.js'
 import en from './locales/en.json'
 import ru from './locales/ru.json'
